LayoutTests/accessibility/table-title.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<title>Table Title</title>
</head>
<body>

<table border="1" id="table">
<caption>High Scores</caption>
<tr>
<th>Score</th>
<th>Name</th>
</tr>
<tr>
<td>9999</td>
<td>Bill</td>
</tr>
<tr>
<td>9998</td>
<td>Ted</td>
</tr>
</table>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that table AXTitle is exposed correctly.");
    
    if (window.accessibilityController) {
        var table = accessibilityController.accessibleElementById("table");
        
        shouldBe("table.title", "'AXTitle: High Scores'");
    }
</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
