Safari-609.1.11.1/LayoutTests/accessibility/table-title-expected.txt

High Scores
Score	Name
9999	Bill
9998	Ted
This tests that table AXTitle is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table.title is 'AXTitle: High Scores'
PASS successfullyParsed is true

TEST COMPLETE
